requirements.txt
boto3==1.34.106
botocore==1.34.109
falcon==3.1.3
gunicorn==22.0.0
jsonschema==4.17.3
requests==2.32.2
SQLAlchemy==2.0.23
urllib3==1.26.18
PyJWT==2.8.0
psycopg2-binary==2.9.5
secure==0.3.0